Adrian's Working version
On my laptop I don't have the couple lines you have added at the beginning of yours, but other than that it's ready. Autocorrelation, 2d graphing and 3d graphing (independently able to turn on or off), searching for best signal for a particular cell through a tetrode, and the ability to decide whether amplitude or sample count is more important.